+++ src/utility.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "utility/algo.h"
+#include "utility/enum.h"
 #include "utility/fs.h"
 #include "utility/io.h"
 #include "utility/string.h"
@@ -8,27 +9,6 @@
 
 namespace mob
 {
-
-// enum stuff
-#define MOB_ENUM_OPERATORS(E) \
-	inline E operator|(E e1, E e2) { return (E)((int)e1 | (int)e2); } \
-	inline E operator&(E e1, E e2) { return (E)((int)e1 & (int)e2); } \
-	inline E operator|=(E& e1, E e2) { e1 = e1 | e2; return e1; }
-
-template <class E>
-bool is_set(E e, E v)
-{
-	static_assert(std::is_enum_v<E>, "this is for enums");
-	return (e & v) == v;
-}
-
-template <class E>
-bool is_any_set(E e, E v)
-{
-	static_assert(std::is_enum_v<E>, "this is for enums");
-	return (e & v) != E(0);
-}
-
 
 // time since mob started
 //
