+++ vs/mob.vcxproj
@@ -133,6 +133,7 @@
     <ClInclude Include="..\src\utility.h" />
     <ClInclude Include="..\src\utility\algo.h" />
     <ClInclude Include="..\src\utility\assert.h" />
+    <ClInclude Include="..\src\utility\enum.h" />
     <ClInclude Include="..\src\utility\fs.h" />
     <ClInclude Include="..\src\utility\io.h" />
     <ClInclude Include="..\src\utility\string.h" />
